pyproject.toml
[tool.poetry]
name = "dstack"
version = "0.1.0"
description = ""
authors = ["Vadim Verein <vadikverein45@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
docker = "^7.0.0"
python-dotenv = "^1.0.1"
boto3 = "^1.34.95"
mypy = "^1.10.0"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110